move marked.js to common
