+++ src/vs/base/common/marked/marked.mock.ts
